+++ examples/mplapack/90_PrecisionComparison/Hilbert_linear_solve_binary80.cpp
@@ -76,7 +76,7 @@ int main() {
     mplapackint *ipiv = new mplapackint[n];
     for (mplapackint j = 0; j < n; j++)
         for (mplapackint i = 0; i < n; i++) {
-            a[i + j * lda] = mplapack_binary80_t(1.0) / mplapack_binary80_t(i + j + 1);
+            a[i + j * lda] = mplapack_binary80_t(1.0) / mplapack_binary80_t((double)(i + j + 1));
             aorg[i + j * lda] = a[i + j * lda];
         }
     for (mplapackint i = 0; i < n; i++)
