+++ examples/mplapack/90_PrecisionComparison/Hilbert_linear_solve_double.cpp
@@ -63,7 +63,7 @@ int main() {
     mplapackint *ipiv = new mplapackint[n];
     for (mplapackint j = 0; j < n; j++)
         for (mplapackint i = 0; i < n; i++) {
-            a[i + j * lda] = double(1.0) / double(i + j + 1);
+            a[i + j * lda] = double(1.0) / double((double)(i + j + 1));
             aorg[i + j * lda] = a[i + j * lda];
         }
     for (mplapackint i = 0; i < n; i++)
