+++ examples/mplapack/90_PrecisionComparison/Hilbert_linear_solve_qd.cpp
@@ -70,7 +70,7 @@ int main() {
     mplapackint *ipiv = new mplapackint[n];
     for (mplapackint j = 0; j < n; j++)
         for (mplapackint i = 0; i < n; i++) {
-            a[i + j * lda] = qd_real(1.0) / qd_real(i + j + 1);
+            a[i + j * lda] = qd_real(1.0) / qd_real((double)(i + j + 1));
             aorg[i + j * lda] = a[i + j * lda];
         }
     for (mplapackint i = 0; i < n; i++)
