+++ examples/mplapack/90_PrecisionComparison/Kahan_rank_reveal_binary128.cpp
@@ -70,7 +70,7 @@ int main() {
     mplapack_binary128_t *aqr = new mplapack_binary128_t[m * n];
     for (mplapackint j = 0; j < n; j++)
         for (mplapackint i = 0; i < m; i++) {
-            mplapack_binary128_t scale = pow(s, mplapack_binary128_t(i));
+            mplapack_binary128_t scale = pow(s, mplapack_binary128_t((double)i));
             mplapack_binary128_t val = (i == j) ? mplapack_binary128_t(1.0) : ((i < j) ? -c : mplapack_binary128_t(0.0));
             a[i + j * lda] = scale * val;
             asvd[i + j * lda] = a[i + j * lda];
